Expose media store wrapper after refactor
