+++ tests/test_create_deck.py
@@ -16,7 +16,7 @@ def test_create_deck_payload(monkeypatch):
         assert params == {"deck": "Inbox"}
         return {"status": "ok"}
 
-    monkeypatch.setattr("anki_mcp.services.anki.anki_call", fake_anki_call)
+    monkeypatch.setattr("anki_mcp.services.client.anki_call", fake_anki_call)
 
     create_deck_fn = deck_helpers._unwrap_tool(create_deck)
     args = CreateDeckArgs(deck="Inbox")
@@ -30,7 +30,7 @@ def test_create_deck_validation_error(monkeypatch):
     async def forbidden_call(action, params):
         raise AssertionError("anki_call should not be invoked on validation errors")
 
-    monkeypatch.setattr("anki_mcp.services.anki.anki_call", forbidden_call)
+    monkeypatch.setattr("anki_mcp.services.client.anki_call", forbidden_call)
 
     create_deck_fn = deck_helpers._unwrap_tool(create_deck)
 
@@ -44,7 +44,7 @@ def test_create_deck_propagates_anki_errors(monkeypatch):
     async def fake_anki_call(action, params):
         raise RuntimeError("boom")
 
-    monkeypatch.setattr("anki_mcp.services.anki.anki_call", fake_anki_call)
+    monkeypatch.setattr("anki_mcp.services.client.anki_call", fake_anki_call)
 
     create_deck_fn = deck_helpers._unwrap_tool(create_deck)
